- add regression tests to ecies_v1 - remove dkg_v0 and ecies_v0, and move the shared code to *_v1
