+++ fastcrypto-tbls/src/dkg_v1.rs
@@ -6,8 +6,9 @@
 //
 
 use crate::dl_verification::verify_poly_evals;
-use crate::nodes::PartyId;
-use crate::polynomial::{Eval, PublicPoly};
+use crate::ecies_v1::{self, RecoveryPackage};
+use crate::nodes::{Nodes, PartyId};
+use crate::polynomial::{Eval, Poly, PrivatePoly, PublicPoly};
 use crate::random_oracle::RandomOracle;
 use crate::tbls::Share;
 use crate::types::ShareIndex;
@@ -17,15 +18,65 @@ use fastcrypto::traits::AllowedRng;
 use itertools::Itertools;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
-
-use crate::dkg::{Complaint, Confirmation, Output, Party};
-use crate::{ecies, ecies_v1};
-
 use tap::prelude::*;
 use tracing::{debug, error, info, warn};
 use zeroize::Zeroize;
 
-// TODO: Move Party, Complaint, Confirmation, Output here and remove old APIs
+/// Generics below use `G: GroupElement' for the group of the VSS public key, and `EG: GroupElement'
+/// for the group of the ECIES public key.
+
+/// Party in the DKG protocol.
+#[derive(Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Party<G: GroupElement, EG: GroupElement>
+where
+    EG::ScalarType: Zeroize,
+{
+    pub id: PartyId,
+    pub(crate) nodes: Nodes<EG>,
+    pub t: u16,
+    pub random_oracle: RandomOracle,
+    pub(crate) enc_sk: ecies_v1::PrivateKey<EG>,
+    pub(crate) vss_sk: PrivatePoly<G>,
+}
+
+/// Assumptions:
+/// - The high-level protocol is responsible for verifying that the 'sender' is correct in the
+///   following messages (based on the chain's authentication).
+/// - The high-level protocol is responsible that all parties see the same order of messages.
+
+/// A complaint/fraud claim against a dealer that created invalid encrypted share.
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Complaint<EG: GroupElement> {
+    pub(crate) accused_sender: PartyId,
+    pub(crate) proof: RecoveryPackage<EG>,
+}
+
+/// A [Confirmation] is sent during the second phase of the protocol. It includes complaints
+/// created by receiver of invalid encrypted shares (if any).
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Confirmation<EG: GroupElement> {
+    pub sender: PartyId,
+    /// List of complaints against other parties. Empty if there are none.
+    pub complaints: Vec<Complaint<EG>>,
+}
+
+// Upper bound on the size of binary serialized incoming messages assuming <=3333 shares, <=400
+// parties, and using G2Element for encryption. This is a safe upper bound since:
+// - Message is O(96*t + 32*n) bytes.
+// - Confirmation is O((96*3 + 32) * k) bytes.
+// Could be used as a sanity safety check before deserializing an incoming message.
+pub const DKG_MESSAGES_MAX_SIZE: usize = 400_000; // 400 KB
+
+/// [Output] is the final output of the DKG protocol in case it runs
+/// successfully. It can be used later with [ThresholdBls], see examples in tests.
+///
+/// If shares is None, the object can only be used for verifying (partial and full) signatures.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Output<G: GroupElement, EG: GroupElement> {
+    pub nodes: Nodes<EG>,
+    pub vss_pk: Poly<G>,
+    pub shares: Option<Vec<Share<G::ScalarType>>>, // None if some shares are missing or weight is zero.
+}
 
 /// [Message] holds all encrypted shares a dealer sends during the first phase of the
 /// protocol.
@@ -110,15 +161,60 @@ where
     /// 2. After *all* parties have sent their ECIES public keys, create the (same) set of nodes.
 
     /// 3. Create a new Party instance with the ECIES private key and the set of nodes.
-    // TODO: Move new() and t() here
+    pub fn new<R: AllowedRng>(
+        enc_sk: ecies_v1::PrivateKey<EG>,
+        nodes: Nodes<EG>,
+        t: u16, // The number of parties that are needed to reconstruct the full key/signature (f+1).
+        random_oracle: RandomOracle, // Should be unique for each invocation, but the same for all parties.
+        rng: &mut R,
+    ) -> FastCryptoResult<Self> {
+        // Confirm that my ecies pk is in the nodes.
+        let enc_pk = ecies_v1::PublicKey::<EG>::from_private_key(&enc_sk);
+        let my_node = nodes
+            .iter()
+            .find(|n| n.pk == enc_pk)
+            .ok_or(FastCryptoError::InvalidInput)?;
+        // Sanity check that the threshold makes sense (t <= n/2 since we later wait for 2t-1).
+        if t > (nodes.total_weight() / 2) || t == 0 {
+            return Err(FastCryptoError::InvalidInput);
+        }
+        // TODO: [comm opt] Instead of generating the polynomial at random, use PRF generated values
+        // to reduce communication.
+        let vss_sk = PrivatePoly::<G>::rand(t - 1, rng);
+
+        // TODO: remove once the protocol is stable since it's a non negligible computation.
+        let vss_pk = vss_sk.commit::<G>();
+        info!(
+            "DKG: Creating party {} with weight {}, nodes hash {:?}, t {}, n {}, ro {:?}, enc pk {:?}, vss pk c0 {:?}",
+            my_node.id,
+            my_node.weight,
+            nodes.hash(),
+            t,
+            nodes.total_weight(),
+            random_oracle,
+            enc_pk,
+            vss_pk.c0(),
+        );
+
+        Ok(Self {
+            id: my_node.id,
+            nodes,
+            t,
+            random_oracle,
+            enc_sk,
+            vss_sk,
+        })
+    }
+
+    /// The threshold needed to reconstruct the full key/signature.
+    pub fn t(&self) -> u16 {
+        self.t
+    }
 
     /// 4. Create the first message to be broadcasted.
     ///
     ///    Returns IgnoredMessage if the party has zero weight (so no need to create a message).
-    pub fn create_message_v1<R: AllowedRng>(
-        &self,
-        rng: &mut R,
-    ) -> FastCryptoResult<Message<G, EG>> {
+    pub fn create_message<R: AllowedRng>(&self, rng: &mut R) -> FastCryptoResult<Message<G, EG>> {
         let node = self.nodes.node_id_to_node(self.id).expect("my id is valid");
         if node.weight == 0 {
             return Err(FastCryptoError::IgnoredMessage);
@@ -169,7 +265,7 @@ where
     }
 
     // Sanity checks that can be done by any party on a received message.
-    fn sanity_check_message_v1(&self, msg: &Message<G, EG>) -> FastCryptoResult<()> {
+    fn sanity_check_message(&self, msg: &Message<G, EG>) -> FastCryptoResult<()> {
         let node = self
             .nodes
             .node_id_to_node(msg.sender)
@@ -244,7 +340,7 @@ where
     ///    and just wait for f+1 valid messages).
     ///
     ///    Assumptions: Called only once per sender (the high level protocol is responsible for deduplication).
-    pub fn process_message_v1<R: AllowedRng>(
+    pub fn process_message<R: AllowedRng>(
         &self,
         message: Message<G, EG>,
         rng: &mut R,
@@ -255,7 +351,7 @@ where
             message.vss_pk.c0()
         );
         // Ignore if invalid (and other honest parties will ignore as well).
-        self.sanity_check_message_v1(&message)?;
+        self.sanity_check_message(&message)?;
 
         let my_share_ids = self.nodes.share_ids_of(self.id).expect("my id is valid");
         let encryption_ro = self.encryption_random_oracle(message.sender);
@@ -340,7 +436,7 @@ where
     ///
     ///    Assumptions: processed_messages is the result of process_message on the same set of messages
     ///    on all parties.
-    pub fn merge_v1(
+    pub fn merge(
         &self,
         processed_messages: &[ProcessedMessage<G, EG>],
     ) -> FastCryptoResult<(Confirmation<EG>, UsedProcessedMessages<G, EG>)> {
@@ -400,7 +496,7 @@ where
     ///    Returns NotEnoughInputs if the threshold minimal_threshold is not met.
     ///
     ///    Assumptions: All parties use the same set of confirmations (and outputs from merge).
-    pub(crate) fn process_confirmations_v1<R: AllowedRng>(
+    pub(crate) fn process_confirmations<R: AllowedRng>(
         &self,
         messages: &UsedProcessedMessages<G, EG>,
         confirmations: &[Confirmation<EG>],
@@ -459,7 +555,7 @@ where
                     .expect("checked above that accuser is valid id");
                 // If the claim refers to a non existing message, it's an invalid complaint.
                 let valid_complaint = match id_to_m1.get(&accused) {
-                    Some(related_m1) => Self::check_complaint_proof_v1(
+                    Some(related_m1) => Self::check_complaint_proof(
                         &complaint.proof,
                         accuser_pk,
                         accuser,
@@ -525,10 +621,7 @@ where
     }
 
     /// 8. Aggregate the valid shares (as returned from the previous step) and the public key.
-    pub(crate) fn aggregate_v1(
-        &self,
-        messages: &VerifiedProcessedMessages<G, EG>,
-    ) -> Output<G, EG> {
+    pub(crate) fn aggregate(&self, messages: &VerifiedProcessedMessages<G, EG>) -> Output<G, EG> {
         debug!(
             "Aggregating shares from {} verified messages",
             messages.0.len()
@@ -599,21 +692,21 @@ where
     }
 
     /// Execute the previous two steps together.
-    pub fn complete_v1<R: AllowedRng>(
+    pub fn complete<R: AllowedRng>(
         &self,
         messages: &UsedProcessedMessages<G, EG>,
         confirmations: &[Confirmation<EG>],
         rng: &mut R,
     ) -> FastCryptoResult<Output<G, EG>> {
-        let verified_messages = self.process_confirmations_v1(messages, confirmations, rng)?;
-        Ok(self.aggregate_v1(&verified_messages))
+        let verified_messages = self.process_confirmations(messages, confirmations, rng)?;
+        Ok(self.aggregate(&verified_messages))
     }
 
     // Returns an error if the *complaint* is invalid (counterintuitive).
     #[allow(clippy::too_many_arguments)]
-    fn check_complaint_proof_v1<R: AllowedRng>(
-        recovery_pkg: &ecies::RecoveryPackage<EG>,
-        receiver_pk: &ecies::PublicKey<EG>,
+    fn check_complaint_proof<R: AllowedRng>(
+        recovery_pkg: &ecies_v1::RecoveryPackage<EG>,
+        receiver_pk: &ecies_v1::PublicKey<EG>,
         receiver_id: PartyId,
         share_ids: &[ShareIndex],
         vss_pk: &PublicPoly<G>,
@@ -672,5 +765,29 @@ where
             "recovery of {} received from {}",
             accuser, accused
         ))
+    }
+}
+
+#[cfg(test)]
+pub fn create_fake_complaint<EG>() -> Complaint<EG>
+where
+    EG: GroupElement + Serialize + DeserializeOwned + HashToGroupElement,
+    <EG as GroupElement>::ScalarType: FiatShamirChallenge + Zeroize,
+{
+    let sk = ecies_v1::PrivateKey::<EG>::new(&mut rand::thread_rng());
+    let pk = ecies_v1::PublicKey::<EG>::from_private_key(&sk);
+    let mr_enc = ecies_v1::MultiRecipientEncryption::encrypt(
+        &[(pk.clone(), b"test".to_vec())],
+        &RandomOracle::new("test"),
+        &mut rand::thread_rng(),
+    );
+    let pkg = mr_enc.create_recovery_package(
+        &sk,
+        &RandomOracle::new("does not matter"),
+        &mut rand::thread_rng(),
+    );
+    Complaint {
+        accused_sender: 1,
+        proof: pkg,
     }
 }
